// my-project/src/app/shared/models/blog.ts
export interface Blog {
    _id?: any;
    name?: any;
    title?: any;
    description?: any;
    smallDescription?: any;
    content?: any;
    aboutQuili?: any;
    aboutPrimitive?: any;
    count?: any;
    thumbnail?: any;
    images?: any;
    createAt?: any;
    socialAccounts?: any;
    category?: any;
    blogType?: any;
    isFeatured?: any;
    meta?: any; 
}